modified menu bar to a common component
